Remove quoted strings from command inputs.
Before this commit, commands were cut whenever a
separator char was encountered. Unfortunately,
when those chars are used as the arguments of a
function, some text is mistaken for a new
command. For me, this manifests itself as URL
chunks being interpreted as commands.

```
[▮▮▮       ]  2.07%  87 ?xt=urn
[▮▮▮       ]  1.81%  76 tr=udp%3A%2F%2F
```

This commits makes every quoted string ignored so
that such false commands no longer happen.
A similar process still needs to be applied for
escaped separators.